Move tahoe hacks into its own file
