+++ search/api.py
@@ -8,6 +8,7 @@ from .filter_generator import SearchFilterGenerator
 from .search_engine_base import SearchEngine
 from .result_processor import SearchResultProcessor
 from .utils import DateRange
+from .tahoe_hacks import has_access_for_results
 
 # Default filters that we support, override using COURSE_DISCOVERY_FILTERS setting if desired
 DEFAULT_FILTER_FIELDS = ["org", "modes", "language"]
@@ -77,48 +78,6 @@ def perform_search(
     return results
 
 
-def _hack_filter_discovery_results(results):
-    """
-    Filter CourseDiscovery search results.
-
-    This is a hack function that should be refactored into the LMS.
-    See RED-637.
-    """
-    if not getattr(settings, 'TAHOE_ENABLE_HAS_ACCESS_FILTER', True):
-        return results
-
-    from lms.djangoapps.courseware.access import has_access  # pylint: disable=import-error
-    from crum import get_current_request  # pylint: disable=import-error
-    from opaque_keys.edx.keys import CourseKey  # pylint: disable=import-error
-
-    user = get_current_request().user
-
-    for result in results["results"]:
-        course_key = CourseKey.from_string(result['data']['id'])
-        if not has_access(user, 'see_in_catalog', course_key):
-            result["data"] = None
-
-    # Count and remove the results that has no access
-    access_denied_count = len([r for r in results["results"] if r["data"] is None])
-    results["access_denied_count"] = access_denied_count
-    results["results"] = [r for r in results["results"] if r["data"] is not None]
-
-    # Hack: Naively reduce the facet numbers by the access denied results
-    # This is not the smartest hack, and customers could report issues
-    # The solution is most likely to just remove the facet numbers
-    results["total"] = max(0, results["total"] - access_denied_count)
-    for _name, facet in list(results["facets"].items()):
-        results["total"] = max(0, results["total"] - access_denied_count)
-        facet["other"] = max(0, facet.get("other", 0) - access_denied_count)
-        facet["terms"] = {
-            term: max(0, count - access_denied_count)
-            for term, count in list(facet["terms"].items())
-            # Remove the facet terms that has no results
-            if max(0, count - access_denied_count)
-        }
-    return results
-
-
 def course_discovery_search(search_term=None, size=20, from_=0, field_dictionary=None):
     """
     Course Discovery activities against the search engine index of course details
@@ -151,4 +110,4 @@ def course_discovery_search(search_term=None, size=20, from_=0, field_dictionary
         facet_terms=course_discovery_facets(),
     )
 
-    return _hack_filter_discovery_results(results)
+    return has_access_for_results(results)
